refactor: deduplicate command setup, remove dead code, consolidate parseFrontmatter
- Extract shared resolveContext/getResourceTypes into cli/context.ts
- Remove duplicate parseFrontmatter from content-block.ts, reuse from utils.ts
- Fix utils.ts parseFrontmatter to use static import instead of require()
- Remove broken catalog apply() "add" branch (empty fields/description) — applyWithLocal handles this correctly
- Remove unused import (join from node:path in catalog.ts)

Net -100 lines of duplicated/dead code.
