+++ src/cli/commands/diff.ts
@@ -1,10 +1,9 @@
 import type { Command } from "commander";
-import { BrazeClient } from "../../core/braze-client.js";
-import { loadConfig, resolveApiKey } from "../../core/config.js";
 import { formatDiffTable } from "../../formatters/table.js";
 import { CatalogProvider } from "../../providers/catalog.js";
 import { ContentBlockProvider } from "../../providers/content-block.js";
 import type { DiffResult } from "../../types/diff.js";
+import { getResourceTypes, resolveContext } from "../context.js";
 
 export function registerDiffCommand(program: Command): void {
 	program
@@ -14,25 +13,9 @@ export function registerDiffCommand(program: Command): void {
 		.option("--resource <type>", "Filter by resource type")
 		.option("--fail-on-drift", "Exit with code 1 if any drift detected")
 		.action(async (opts) => {
-			const configPath = program.opts().config ?? "braze-sync.config.yaml";
-			const verbose = program.opts().verbose ?? false;
-
-			const config = await loadConfig(configPath);
-			const env = config.environments[opts.env];
-			if (!env) {
-				console.error(`Error: Environment '${opts.env}' not found in config`);
-				process.exit(1);
-			}
-
-			const apiKey = resolveApiKey(env.api_key_env);
-			const client = new BrazeClient({ apiUrl: env.api_url, apiKey, verbose });
-
+			const { config, client } = await resolveContext(program, opts.env);
 			const allDiffs: DiffResult[] = [];
-			const resourceTypes = opts.resource
-				? [opts.resource]
-				: Object.keys(config.resources).filter(
-						(k) => config.resources[k as keyof typeof config.resources],
-					);
+			const resourceTypes = getResourceTypes(config, opts.resource);
 
 			for (const resourceType of resourceTypes) {
 				if (resourceType === "catalogs" && config.resources.catalogs) {
